Chat With multiple Pdf Documents with Langchain and Google Gemini Pro/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain_community
PyPDF2
chromadb
faiss-cpu
langchain-google-genai